refactor: apply PR feedback
