+++ test-utils/src/index.d.ts
@@ -1,4 +1,3 @@
 export function setupRerender(): () => void;
 export function act(callback: () => void): void;
-export function setupScratch(): HTMLDivElement;
 export function teardown(): void;
