+++ test-utils/test/shared/teardown.test.js
@@ -1,5 +1,6 @@
 const { Component, render } = require('preact');
-const { teardown, setupRerender, setupScratch } = require('../../src');
+const { teardown, setupRerender } = require('../../src');
+const { setupScratch } = require('../../../test/_util/helpers');
 
 describe('teardown', () => {
 	let rerender, scratch;
@@ -7,15 +8,6 @@ describe('teardown', () => {
 		rerender = setupRerender();
 		scratch = setupScratch();
 	});
-
-	it('should clear the children of scratch', () => {
-		render('<div>inner</div>', scratch);
-		teardown(scratch);
-		setTimeout(() => {
-			expect(scratch.firstChild).to.be.null;
-			expect(scratch.innerHTML).to.equal('');
-		}, 0);
-	})
 
 	it('should restore debounce', () => {
 		teardown(scratch);
